Rename to Arm Toolchain for Embedded and update directories
This updates all the files with the toolchain's new name and
directory structure. The main functional difference is that because
the CMake scripts now live inside a fork of llvm-project, it is no
longer necessary to checkout that repo. However the downstream
patches remain for the moment as these have not yet been applied to
the new repo, and will need to be manually applied before a build.

Documentation will likely need further updates, but these should
be enough to start building.