The ATfL build script (#15)
This is just a build script, it does not handle system (docker)
configuration or build-time dependencies (build essentials) which are
assumed to be already available. Also it just creates a tarball, not the
deb or rpm neither it enforces ArmPL dependency, although it assumes a
path to the ArmPL's libamath directory is being provided.

Read --help message for more details.