[automerge] Use separate git restore commands for automerge_ignore (#21)
The latest version of git, which is used in GitHub Actions, does not
allow the use of the `--ours` and `--staged` options together in the
same `git restore` command. This splits the command into two calls, one
using `--staged` to restore ignored files in the index and one using
`--ours --worktree` to restore ignored files in the working tree.