Add arm-toolchain CONTRIBUTING.md including downstream patch policy (#48)
This is based on the "LLVM Embedded Toolchain for Arm" contributing.md
with adaptations for the arm-toolchain.

The largest change is the introduction of the downstream patch policy.
Arm wants the arm-toolchain to be as close to the upstream open-source
projects as possible, we impose some additional requirements on
downstream patches so that we can summarize all the patches and know why
they were made.

Replace llvm-project's CONTRIBUTING.md rather than using another
location like .github as we expect many developers to read the file from
the terminal.

As a result of the downstream patch policy, introduce a docs folder in
Arm Software to contain documentation common to all toolchains. Include
some skeleton toolchain feature markdown files ready to edit.

Remove the old LLVM ET contributing.md as this supersedes it.